pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codebase-cartographer"
version = "0.1.0"
description = "A tool for mapping and analyzing codebases"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Aman", email = "aman@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "networkx>=3.0",
    "pydantic>=2.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-python",
    "tree-sitter-javascript",
    "tree-sitter-yaml",
    "tree-sitter-sql",
    "sqlglot>=20.0.0",
    "scipy>=1.8.0",
    "numpy>=1.21.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
codebase-cartographer = "cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
